feat: modular agent execution
